refactor: 변수 네임을 가독성이 좋은 형태로 변경
