+++ src/appConfig.ts
@@ -1,0 +1,2 @@
+export const API_BASE_URL = 'https://shopping-cart.techcourse.co.kr';
+export const CUSTOMER_NAME = '0307kwon';
