--- src/constants/API.ts
@@ -1,7 +1,0 @@
-//TODO: 이 파일은 config로 빠져야 되지 않을까?
-
-// JSON Server
-export const API_BASE_URL = 'https://shopping-cart.techcourse.co.kr';
-export const CUSTOMER_NAME = '0307kwon';
-
-export type HTTPMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
