+++ src/hooks/useCartItems/useCartChangeCheckState/index.ts
@@ -4,7 +4,7 @@ import { CartItem } from '../../../types';
 
 const useCartChangeCheckState = () => {
   const dispatch = useAppDispatch();
-  const [itemsInCart] = useAppSelector(({ cart: { items, error, isLoading } }) => [
+  const [cartItems] = useAppSelector(({ cart: { items, error, isLoading } }) => [
     items,
     error,
     isLoading,
@@ -15,7 +15,7 @@ const useCartChangeCheckState = () => {
   };
 
   const changeAllChecked = (checked: boolean) => {
-    dispatch(thunkChangeAllItemChecked(itemsInCart, checked));
+    dispatch(thunkChangeAllItemChecked(cartItems, checked));
   };
 
   return { toggleChecked, changeAllChecked };
